epiar: moved to cvs and now it is both 64 and 32 bit yay :) and some fixes to bjs
I moved epiar to cvs, despite the fact that its only a tech demo and less
playable than the old one. the reasons are simple 1) it looks freaking awsome!
2) it works on both 64 and 32, and i try to make sure that all free packages
will be compatible on both 3) no one used epiar 0.5.0 for ages, no one will
mess it, old epiar is ugly, not stable, not playable, and wouldnt be here for
to long, if it was at least popular, or good then i would have kept it..

bjs got some small fixes (md5sum addon...)

and i added a not for myself in todo

Signed-off-by: vleon <vleon1@gmail.com>